EE: use (u64)(i16) instead of (i32)(i16) for sign extension

+++ src/EmotionEngine/x64/opcodes_x64.cpp
@@ -139,12 +139,12 @@ void JitX64::MTC0(InstructionData& data) {
 
 void JitX64::ADDIU(InstructionData& data) {
 	if (data.rt == 0) return;
-	cc.lea(r[data.rt].r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(r[data.rt].r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	cc.movsxd(r[data.rt], r[data.rt].r32());
 }
 
 void JitX64::SW(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	EmitWriteVirtualMemory<u32>(temp.r32(), r[data.rt].r32());
 }
 
@@ -155,7 +155,7 @@ void JitX64::JALR(InstructionData& data) {
 }
 
 void JitX64::SD(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	EmitWriteVirtualMemory<u64>(temp.r32(), r[data.rt]);
 }
 
@@ -192,7 +192,7 @@ void JitX64::BGEZ(InstructionData& data) {
 }
 
 void JitX64::LBU(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	if (data.rt == 0) {
 		EmitReadVirtualMemory<u8>(temp.r8(), temp.r32());
 		return;
@@ -229,7 +229,7 @@ void JitX64::BEQ(InstructionData& data) {
 }
 
 void JitX64::LD(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	if (data.rt == 0) {
 		EmitReadVirtualMemory<u64>(temp.r64(), temp.r32());
 		return;
@@ -274,7 +274,7 @@ void JitX64::DADDU(InstructionData& data) {
 }
 
 void JitX64::LW(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	if (data.rt == 0) {
 		EmitReadVirtualMemory<u32>(temp.r32(), temp.r32());
 		return;
@@ -369,7 +369,7 @@ void JitX64::SLTIU(InstructionData& data) {
 }
 
 void JitX64::LB(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	if (data.rt == 0) {
 		EmitReadVirtualMemory<u8>(temp.r8(), temp.r32());
 		return;
@@ -381,7 +381,7 @@ void JitX64::LB(InstructionData& data) {
 
 void JitX64::SWC1(InstructionData& data) {
 	x86::Gp address = cc.new_gp32();
-	cc.lea(address, x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(address, x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	cc.mov(temp.r32(), x86::dword_ptr(r5900, offsetof(R5900, fpr) + (sizeof(float) * data.rt)));
 	EmitWriteVirtualMemory<u32>(address, temp.r32());
 }
@@ -392,7 +392,7 @@ void JitX64::J(InstructionData& data) {
 }
 
 void JitX64::SB(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	EmitWriteVirtualMemory<u8>(temp.r32(), r[data.rt].r8());
 }
 
@@ -474,7 +474,7 @@ void JitX64::DSRL32(InstructionData& data) {
 }
 
 void JitX64::LHU(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	if (data.rt == 0) {
 		EmitReadVirtualMemory<u16>(temp.r16(), temp.r32());
 		return;
@@ -503,7 +503,7 @@ void JitX64::BLTZ(InstructionData& data) {
 }
 
 void JitX64::SH(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	EmitWriteVirtualMemory<u16>(temp.r32(), r[data.rt].r16());
 }
 
@@ -523,7 +523,7 @@ void JitX64::XORI(InstructionData& data) {
 }
 
 void JitX64::LWU(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	if (data.rt == 0) {
 		EmitReadVirtualMemory<u32>(temp.r32(), temp.r32());
 		return;
@@ -552,11 +552,11 @@ void JitX64::DSLLV(InstructionData& data) {
 
 void JitX64::DADDIU(InstructionData& data) {
 	if (data.rt == 0) return;
-	cc.lea(r[data.rt], x86::ptr(r[data.rs], (i32)(i16)data.imm));
+	cc.lea(r[data.rt], x86::ptr(r[data.rs], (u64)(i16)data.imm));
 }
 
 void JitX64::LH(InstructionData& data) {
-	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (i32)(i16)data.imm));
+	cc.lea(temp.r32(), x86::ptr(r[data.rs].r32(), (u64)(i16)data.imm));
 	if (data.rt == 0) {
 		EmitReadVirtualMemory<u16>(temp.r16(), temp.r32());
 		return;
